【缩放】功能完成；修复平移、旋转的bug
平移Bug：之前的平移只改了点集而没有变参数，所以利用虚函数机制重构了平移功能；
旋转Bug：getRotateR()中 y不小心写成了x（本应该是"-yr"）

+++ CG_Painter/CG_Painter/canvas.cpp
@@ -102,7 +102,7 @@ void Canvas::rotate(int id, int x, int y, int r)
 	}
 }
 
-void Canvas::scale(int id, int x, int y, float s)
+void Canvas::scale(int id, int x, int y, double s)
 {
 	for (auto i = pixelsets.begin(); i != pixelsets.end(); i++) {
 		if ((*i)->id == id) {
